## Premise:
I've been using this library to compare very large hashes. The test hashes I used were both built from 10k+ lines of YAML and I found calculating the diff between the two hashes took around 90 seconds (304 seconds when profiling). I used [ruby-prof](https://github.com/ruby-prof/ruby-prof) to try and find some room for improvement, and managed to find a couple of easy wins.

## Changes:
- Instead of checking `is_a?`, extend `Object`, `Array` and `Hash` with their own `count_nodes` methods
  - This dropped the total time from 304 seconds to around 270
- Cache the count of the nodes in `Hash` objects 
  - This dropped the total time from 270 to about 197
  - There wasn't much to be gained from caching the count for `Array` objects
  - I think this is safe - from my understanding of the code, it does not add/remove objects from the original hashes so the count would not change after the first calculation

## Concerns:
Although this is a pretty significant time savings (for me at least - without profiling I dropped the calculation time from 90 seconds to about 65 seconds), extending/modifying classes and the original objects may be undesirable for whatever reason. I leave this up to your judgment @liufengyun.