@liufengyun 2016-04-05
Thanks @kgooble , it's nice you've run the program against a large data set.

I definitely welcome performance improvement. In this specific case, I think the performance gain is not significant enough to justify messing with `Array`, `Hash` and `Object`. If the performance gain is say several times, I'll merge the PR without hesitation.